+++ meteor-packages/mdc-sass/package.js
@@ -1,18 +1,30 @@
-var config = {
-  "name": "zodiase:mdc-sass",
-  "version": "0.0.0_3",
-  "summary": "Material Components for Meteor (SASS only)",
-  "git": "https://github.com/Zodiase/meteor-mdc.git",
-  "meteorRelease": "1.4.2.3",
-  "scssVersion": '3.13.0',
-  "sassFiles": ["./bundle.scss","./animation/_functions.scss","./animation/_mixins.scss","./animation/_variables.scss","./animation/mdc-animation.scss","./button/mdc-button.scss","./card/mdc-card.scss","./checkbox/_keyframes.scss","./checkbox/_variables.scss","./checkbox/mdc-checkbox.scss","./drawer/_mixins.scss","./drawer/mdc-drawer.scss","./drawer/permanent/mdc-permanent-drawer.scss","./drawer/temporary/mdc-temporary-drawer.scss","./elevation/_mixins.scss","./elevation/_variables.scss","./elevation/mdc-elevation.scss","./fab/mdc-fab.scss","./form-field/mdc-form-field.scss","./icon-toggle/mdc-icon-toggle.scss","./list/mdc-list.scss","./menu/mdc-menu.scss","./menu/simple/mdc-simple-menu.scss","./radio/mdc-radio.scss","./ripple/_keyframes.scss","./ripple/_mixins.scss","./ripple/_variables.scss","./ripple/mdc-ripple.scss","./rtl/_mixins.scss","./select/mdc-select.scss","./snackbar/_variables.scss","./snackbar/mdc-snackbar.scss","./textfield/mdc-textfield.scss","./theme/_constants.scss","./theme/_functions.scss","./theme/_mixins.scss","./theme/_variables.scss","./theme/mdc-theme.scss","./typography/_mixins.scss","./typography/_variables.scss","./typography/mdc-typography.scss"] // Generated
-};
+//  Copyright 2017 Xingchen Hong
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+
+var loadJSON = function (fs) {
+  return function (uri) {
+    return JSON.parse(fs.readFileSync(uri));
+  };
+}(Npm.require('fs'));
+
+var config = loadJSON('./package.json');
 
 Package.describe({
-  name: config.name,
+  name: config.meteorPackageName,
   version: config.version,
   // Brief, one-line summary of the package.
-  summary: config.summary,
+  summary: config.description,
   // URL to the Git repository containing the source code for this package.
   git: config.git,
   // By default, Meteor will default to using README.md for documentation.
@@ -27,7 +39,7 @@ Package.onUse(function(api) {
 
   // Add sass files to be imported by users.
   // Has to be added to `client` platform or `fourseven:scss` won't see them.
-  api.addFiles(config.sassFiles, 'client', {isImport: true});
+  api.addFiles(config._sassFiles, 'client', {isImport: true});
 
   api.mainModule('main.js', 'client');
 });
@@ -35,6 +47,9 @@ Package.onUse(function(api) {
 Package.onTest(function(api) {
   api.use('ecmascript');
   api.use('tinytest');
-  api.use(config.name);
+  // Use the SCSS package to enable importing SASS files.
+  api.use('fourseven:scss@' + config.scssVersion);
+  api.use(config.meteorPackageName);
+
   api.mainModule('main-tests.js', 'client');
 });