+++ meteor-packages/mdc/package.js
@@ -12,24 +12,19 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-var config = {
-  "name": "zodiase:mdc",
-  "version": "0.0.2",
-  "summary": "Material Components for Meteor",
-  "git": "https://github.com/Zodiase/meteor-mdc.git",
-  "jsVersion": "0.0.1",
-  "cssVersion": "0.0.2",
-  "meteorRelease": "1.4.2.3",
-  "clientAssets": [
-    'stylesheet.js'
-  ]
-};
+var loadJSON = function (fs) {
+  return function (uri) {
+    return JSON.parse(fs.readFileSync(uri));
+  };
+}(Npm.require('fs'));
+
+var config = loadJSON('./package.json');
 
 Package.describe({
-  name: config.name,
+  name: config.meteorPackageName,
   version: config.version,
   // Brief, one-line summary of the package.
-  summary: config.summary,
+  summary: config.description,
   // URL to the Git repository containing the source code for this package.
   git: config.git,
   // By default, Meteor will default to using README.md for documentation.
@@ -43,7 +38,9 @@ Package.onUse(function(api) {
   api.use('zodiase:mdc-styleless@=' + config.jsVersion);
   api.use('zodiase:mdc-styleonly@=' + config.cssVersion);
 
-  api.addAssets(config.clientAssets, 'client');
+  api.addAssets([
+    'stylesheet.js'
+  ], 'client');
 
   api.mainModule('main.js', 'client');
 });
@@ -51,6 +48,7 @@ Package.onUse(function(api) {
 Package.onTest(function(api) {
   api.use('ecmascript');
   api.use('tinytest');
-  api.use(config.name);
+  api.use(config.meteorPackageName);
+
   api.mainModule('main-tests.js', 'client');
 });