+++ meteor-packages/mdc-styleless/CHANGELOG.md
@@ -1,4 +1,6 @@
 Change Log
 ------------------------------------------------------------------------------
+* v0.0.2
+    * Update to MDC 0.4.0.
 * v0.0.1
     * Serves the bundle JS file from the Material Components NPM package.