+++ meteor-packages/mdc-styleonly/CHANGELOG.md
@@ -1,5 +1,7 @@
 Change Log
 ------------------------------------------------------------------------------
+* v0.0.3
+    * Update to MDC 0.4.0.
 * v0.0.2
     * Now the CSS file is added as an asset so it won't load automatically. This allows the users to better control when the stylesheet should be loaded.
 * v0.0.1