+++ meteor-packages/mdc/CHANGELOG.md
@@ -1,5 +1,9 @@
 Change Log
 ------------------------------------------------------------------------------
+* v0.0.3
+    * Uses `zodiase:mdc-styleless@0.0.2`.
+    * Uses `zodiase:mdc-styleonly@0.0.3`.
+    * (The above means update NPM Material Components to 0.4.0).
 * v0.0.2
     * Uses `zodiase:mdc-styleonly@0.0.2`.
     * Adds a shorthand for loading the stylesheet.